Moved examples into their own directories.
