+++ src/examples/CMakeLists.txt
@@ -1,4 +1,6 @@
 CMAKE_MINIMUM_REQUIRED(VERSION 3.0)
 
-ADD_SUBDIRECTORY(black_boxes)
 ADD_SUBDIRECTORY(chat)
+ADD_SUBDIRECTORY(key_value_store)
+ADD_SUBDIRECTORY(primality)
+ADD_SUBDIRECTORY(redis)
