--- src/examples/black_boxes/CMakeLists.txt
@@ -1,17 +1,0 @@
-CMAKE_MINIMUM_REQUIRED(VERSION 3.0)
-
-MACRO(CREATE_BLACK_BOX_EXECUTABLE NAME)
-    ADD_EXECUTABLE(examples_black_boxes_${NAME} ${NAME})
-    TARGET_LINK_LIBRARIES(examples_black_boxes_${NAME} fluent)
-    ADD_DEPENDENCIES(examples_black_boxes_${NAME} fluent)
-ENDMACRO(CREATE_BLACK_BOX_EXECUTABLE)
-
-CREATE_BLACK_BOX_EXECUTABLE(key_value_client)
-CREATE_BLACK_BOX_EXECUTABLE(key_value_server)
-CREATE_BLACK_BOX_EXECUTABLE(primality_client)
-CREATE_BLACK_BOX_EXECUTABLE(primality_server)
-CREATE_BLACK_BOX_EXECUTABLE(redis_client)
-CREATE_BLACK_BOX_EXECUTABLE(redis_server)
-
-TARGET_LINK_LIBRARIES(examples_black_boxes_redis_server redox ev hiredis)
-ADD_DEPENDENCIES(examples_black_boxes_redis_server redoxlib)
