+++ src/examples/key_value_store/CMakeLists.txt
@@ -1,0 +1,10 @@
+CMAKE_MINIMUM_REQUIRED(VERSION 3.0)
+
+MACRO(CREATE_KEY_VALUE_STORE_EXECUTABLE NAME)
+    ADD_EXECUTABLE(examples_key_value_store_${NAME} ${NAME})
+    TARGET_LINK_LIBRARIES(examples_key_value_store_${NAME} fluent)
+    ADD_DEPENDENCIES(examples_key_value_store_${NAME} fluent)
+ENDMACRO(CREATE_KEY_VALUE_STORE_EXECUTABLE)
+
+CREATE_KEY_VALUE_STORE_EXECUTABLE(key_value_client)
+CREATE_KEY_VALUE_STORE_EXECUTABLE(key_value_server)
