+++ src/examples/key_value_store/README.md
@@ -1,0 +1,17 @@
+# Key-Value Store
+
+```bash
+# Configuration
+DB_USER="username"
+DB_PASS="password"
+DB_NAME="the_db"
+SERVER_ADDR="tcp://0.0.0.0:9000"
+CLIENT_ADDR="tcp://0.0.0.0:9001"
+REDIS_ADDR=localhost
+REDIS_PORT=6379
+
+GLOG_logtostderr=1 ./build/examples_key_value_store_key_value_server \
+    $DB_USER $DB_PASS $DB_NAME $SERVER_ADDR
+GLOG_logtostderr=1 ./build/examples_key_value_store_key_value_client \
+    $DB_USER $DB_PASS $DB_NAME $SERVER_ADDR $CLIENT_ADDR
+```
