+++ src/examples/key_value_store/launch_key_value_store.sh
