+++ src/examples/primality/CMakeLists.txt
@@ -1,0 +1,10 @@
+CMAKE_MINIMUM_REQUIRED(VERSION 3.0)
+
+MACRO(CREATE_PRIMALITY_EXECUTABLE NAME)
+    ADD_EXECUTABLE(examples_primality_${NAME} ${NAME})
+    TARGET_LINK_LIBRARIES(examples_primality_${NAME} fluent)
+    ADD_DEPENDENCIES(examples_primality_${NAME} fluent)
+ENDMACRO(CREATE_PRIMALITY_EXECUTABLE)
+
+CREATE_PRIMALITY_EXECUTABLE(primality_client)
+CREATE_PRIMALITY_EXECUTABLE(primality_server)
