+++ src/examples/redis/README.md
@@ -1,4 +1,4 @@
-# Black Boxes
+# Redis
 
 ```bash
 # Configuration
@@ -10,19 +10,6 @@ CLIENT_ADDR="tcp://0.0.0.0:9001"
 REDIS_ADDR=localhost
 REDIS_PORT=6379
 
-# Primality.
-GLOG_logtostderr=1 ./build/examples_black_boxes_primality_server \
-    $DB_USER $DB_PASS $DB_NAME $SERVER_ADDR
-GLOG_logtostderr=1 ./build/examples_black_boxes_primality_client \
-    $DB_USER $DB_PASS $DB_NAME $SERVER_ADDR $CLIENT_ADDR
-
-# Key Value.
-GLOG_logtostderr=1 ./build/examples_black_boxes_key_value_server \
-    $DB_USER $DB_PASS $DB_NAME $SERVER_ADDR
-GLOG_logtostderr=1 ./build/examples_black_boxes_key_value_client \
-    $DB_USER $DB_PASS $DB_NAME $SERVER_ADDR $CLIENT_ADDR
-
-# Redis.
 GLOG_logtostderr=1 ./build/examples_black_boxes_redis_server \
     $DB_USER $DB_PASS $DB_NAME $REDIS_ADDR $REDIS_PORT $SERVER_ADDR
 GLOG_logtostderr=1 ./build/examples_black_boxes_redis_client \
